Move components into directory
Clean up typed name states
