+++ src/components/Header.js
@@ -2,11 +2,15 @@ import React, { Component } from 'react';
 import Typist from 'react-typist';
 import './Header.css';
 
-// Typed name states
-const INIT_NICKNAME = 0;
-const FULL_NAME = 1;
-const NICKNAME = 2;
+// Name & related states
+const NICKNAME = 'Will';
+const REST_OF_NAME = 'iam Lorey';
 
+const STATIC_NICKNAME = 0;
+const TYPED_FULL_NAME = 1;
+const TYPED_NICKNAME = 2;
+
+// Cursor for TypedName
 const cursor = {
   show: true,
   blink: false,
@@ -16,35 +20,35 @@ const cursor = {
 };
 
 function TypedName(props) {
-  if (props.nameState === INIT_NICKNAME) {
-    return <Typist cursor={cursor}>Will</Typist>;
-  } else if (props.nameState === FULL_NAME) {
+  if (props.nameState === STATIC_NICKNAME) {
+    return <div>{NICKNAME}</div>;
+  } else if (props.nameState === TYPED_FULL_NAME) {
     return (
       <div>
-        Will
+        {NICKNAME}
         <Typist cursor={cursor} onTypingDone={props.onTypingDone}>
-          iam Lorey
-          <Typist.Backspace count={9} delay={800} />
+          {REST_OF_NAME}
+          <Typist.Backspace count={REST_OF_NAME.length} delay={800} />
         </Typist>
       </div>
     );
   } else {
-    // props.nameState === NICKNAME
-    return <div>Will</div>;
+    // props.nameState === TYPED_NICKNAME
+    return <Typist cursor={cursor}>{NICKNAME}</Typist>;
   }
 }
 
 export default class Header extends Component {
   state = {
-    nameState: INIT_NICKNAME,
+    nameState: TYPED_NICKNAME,
   };
 
   expandName = () => {
-    this.setState({ nameState: FULL_NAME });
+    this.setState({ nameState: TYPED_FULL_NAME });
   };
 
   shrinkName = () => {
-    this.setState({ nameState: NICKNAME });
+    this.setState({ nameState: STATIC_NICKNAME });
   };
 
   render() {
